create contants.js
